"fixed" default credits system for my needs
This will change the system just a bit to the "old" one. If there are more then 1 Traitors, every traitor will get the number of credits set with "ttt_credits_starting" minus the amount of OTHER (the one getting the credits not included) traitors.

Example:

ttt_credits_starting = 3
Current round there are 2 traitors.

So math:
ttt_credits_starting - 1 = 2 . So everyone will get 2 credits...

The formula is like this:

ttt_credits_starting - (AMOUNTOFTRAITORS-1)

Oh and of course if the credits would go to 0 or under, it will reset to 1. So all traitors will get at least 1 credit at start.