Revert "Set gmod sprint speed..." | Add IG to buggy addons
This reverts the change made with the commit f1fb87a526e7929581c72b5b5eceb57993fef98f

as this is already done in the Player:InitialSpawn() method and thus is redundant.

And this also adds the Infinity Gauntlet SEWP (https://steamcommunity.com/sharedfiles/filedetails/?id=1382102057) as a buggy addon, as this is not intended for use with TTT and sets the run speed to 400 on player spawn, which will bring back a default gmod sprint etc.