Moved legacy item init to items module (#584)
In this change, the function to initialize legacy items is now properly contained inside the items module and is called directly in the respective InitPostEntity functions on the client / server, instead of using the hook. 

This also removes the timer delay for migration messages, as there should not be any difference.