Fixed role sync in respawn-cases
This fixed the following issue:
If a player (for example a Traitor) was confirmed and then respawned, he would then show as Innocent as soon as a SendFullState update is triggered (which can happen quite often during a round).
Also I changed the reset from EndRound to PrepareRound because you can still confirm bodys after the round has ended.