Fixed Subroles not being affected by Forced Roles
The Upgrade for subroles now checks if the role is still available by deducting forced roles.
This also enables these functions to be reused without affecting the global selectable roles in any way.
As a compromise the efficiency is decreased as the functions have to always check and calculate the amount of deducted roles.