Fixed several dependency-Issues
Depending on the order, that roles are forced in, it could come to some problems. Therefore we check now, if the the baseroles has enough roles after forcing, available to the subrole and also consider subroles, that have taken place of baseroles.

Order seems to depend on number of forces, or is it random?