remove shared migration files
There might be cases where you want to have the same table on the server and on the client, but we would run into problems if they would diverge at some point. So explicitly setting a migration on the client and the server side respectively should be the way to go.