Move shop favorites functions to cl_equip & remove its autorun file
This patch moves the favorite functions to the clientside shop file and makes them local functions. This is a breaking change and those functions are now not available globally anymore. The convars are also moved to the serverside shop file and are no longer needed to be loaded via an extra autorun file.