Cleanup library structure (#821)
* Move bind library to ttt2/libraries folder

This commit moves the autorun included `bind` library to the libraries folder and removes the need for an additional autorun load file.

* Migrate modules to libraries folder

This patch moves the `huds`, `hudelements`, `pon` and `items` modules to the libraries folder for consistency.

* Move shop favorites functions to cl_equip & remove its autorun file

This patch moves the favorite functions to the clientside shop file and makes them local functions. This is a breaking change and those functions are now not available globally anymore. The convars are also moved to the serverside shop file and are no longer needed to be loaded via an extra autorun file.