consistent text positioning in pure_skin bar
I noticed, that the text positioning in a pure_skin bar is not really consistent at different screen resolutions.
This has caused problems especially with a new HUD element, but it is also noticeable in the player info panel. 
I have tested this at various screen resolutions between 720p and 3000x2000, and it seems to be positioned consistently with the proposed change. 
It also looks good with the Vampire and Bodyguard role. But I have not tested more than that.