Fix missing ids
If you register weapons manually before TTT2 is loaded, it can happen, that those weapons dont get any ids necessary for the shop. Those are the minority as normal weapons are going over the tttweapon_base anyways and need to be initialized in a terrortown folder, therefore wouldnt cause any issues.

This should fix the error of PAC3 https://steamcommunity.com/sharedfiles/filedetails/?id=104691717 with the PAC3\lua\pac3\extra\shared\hands.lua file. Or anything that does the same manual registering in a file-loading with weapons.Register(SWEP).