Fix available baseroles calculation: Consider maxPlys when filtering
This patch potentially corrects the filtering of available baseroles.
When the system goes through all layers for each baserole it does not consider the amount of players that can take the role and thus how many players can already be assigned to a role.
With the new condition it will now correctly stop adding baseroles to the final baserole selection table (marked as available), if the amount is equal to the player amount. This ensures that layering actually influences the availability of base roles and thus wont lead to weird issues further down.
The final role assignment also implicitly assumes that the actual selection of what roles appear is already made in the filtering of available roles. This should now work as intended.

This patch still needs testing and verification.