CD: Add primitive continuous deployment workflow
For now this pretty basic and makes a few assumptions on the environment
target where this should be deployed to:

- correctly configured secrets of the github environment
- already installed and running garrysmod server on the remote host
- correctly configured ssh key on the remote host

Additionally manual intervention on the garrysmod server might be
required if hotreloading breaks.

Besides these points this specific deployment is mainly meant for having
a quick look at a specific pull request on a server and manual testing
the changes on a dedicated resource.