playermodels: Add option to disable resetting player color via `ttt_enforce_playercolor` convar
This is mostly done to enhance compatibility with colourable
playermodels.

Before this change coloured playermodels would be set to `COLOR_WHITE´ or
`GAMEMODE.playercolor` on (re)spawn to prevent players from being
distinguished by different colours.

After this change this behaviour is not changed unless `ttt_enforce_playercolor` is set to 0.
In that case we return early in the `GM:TTTPlayerSetColor` hook and do
nothing.
This does NOT however prevent other addons from hooking into the `TTTPlayerSetColor` hook nor
does it prevent other addons from setting the playercolor somewhere else.