chore: Add .luarc.json to gitignore and add example
The example is missing the `workspace.library` as i haven't found a
proper way to integrate this with the glua-api-snippets [1] yet in my
Neovim setup. In my current setup `workspace.library` points to a local
clone.

VSCode Users should be able to use the `Lua Language Server Addon Manager`
to enable the garrysmod addon (if they have the lua language server
installed).
I haven't tested this, therefore i don't know if the .luarc.json is even
necessary in that case.

This file is mostly set here for reference purposes.