fix: Correct network var names
I failed to see that with the old `AccessorFuncDT()` call it was
possible to influence the accessor function names which resulted in
accessors that were cased differently than their respective networked
vars.

This renames the networked vars which shouldn't itself matter as the
vars should only ever be accessed via their accessors anyway.