fix: Remove usage of `DTVar()`
This commit removes the last usages of `DTVar()` in our code and
replaces them with `NetworkVar()`.

This also replaces direct access to `self.dt.var` with their respective
accessors.

Another thing to mention is that zm_carry was using two bools in the
same "slot". I have no idea what the implications of this were. It is
fixed now.