fix: Deprecate `AccessorFuncDT` instead of removing it (#1123)
Seems like a few other addons are using this function and removing it
causes them to break. So we will deprecate it instead and give addon
developers a chance to update their addons.

Sadly i have not found a good way to log which addon is still using
this. Using `ErrorNoHaltWithStack()` will show the addon in the
stacktrace but it will also be presented to players as "The addon 'TTT2'
is causing errors".
So i went with `MsgN()` for now.