fix: prevent nil error in database module
In cases where `itemName` is not in the `accessName` database table a
nil error would occur while trying to index `sqlData`.

`sqlData` is populated via the ORM library `:find(itemName)` function
which returns the database entry with `itemName` as its primary key.
This returns nil if no such entry is found.

This behaviour was not checked here and results in breaking the
equipment shop.

The minimal working fix would be to if-clause only the `sqlData[_key]`
access. I've scoped the if-clause over a few more items which should
all be NOOPs in case of `sqlData` being nil.