add(fastutf8): Faster utf8 library based on regex matching
From https://github.com/blitmap/lua-utf8-simple

This is much faster and allows for this to be used as is, instead of needing to rewrite the UI to not layout the text on every frame.

But there is no guarantee that this works in every case, as we did not have the time to look into the utf8 spec in detail. Although it did very well in our different non-utf8 languages.