Added Custom World and View Model Rendering (#1325)
The [SWEP Construction
Kit](https://steamcommunity.com/sharedfiles/filedetails/?id=109724869)
is a popular addon that allows the creation of custom view and world
models. Addons such as the boomerang and the melon mine use it.
I liked the core functionality so much, that I refactored the code and
added it to TTT2. The toolkit has more features than this
implementation, but the core functionality of adding world and
viewmodels is there. The data tables are also compatible with those
created by the toolkit.


![image](https://github.com/TTT-2/TTT2/assets/13639408/eacd10eb-7f80-4154-ae56-b655c89a3759)
---------

Co-authored-by: saibotk <git@saibotk.de>
Co-authored-by: Histalek <16392835+Histalek@users.noreply.github.com>