MarkerVision: Added Static Role/Team Support and API Rework (#1336)
While reworking TTT2 Heroes a bit I noticed that we need markers that
are not only bound to a player, but to a role or team. Therefore I
extended the code so that the owner can be a player or a role/team.
While doing this I noticed a few other issues with the markerVision
library, most notably that is was impossible to have two marks (for
different players) on the same identity. This pullequest resolves this.

E.g. the heroes crystal now works like this:

```lua
local mvObject = self:AddMarkerVision("crystal_supervillain")
mvObject:SetOwner(ROLE_SUPERVILLAIN)
mvObject:SetVisibleFor(VISIBLE_FOR_ROLE)
mvObject:SyncToClients()
```

This also already tracks role changes. So when a player changes their
role, they lose access to the wallhack. But they can gain it back, if
they change their role back.

## Note

If you'd replace `ROLE_SUPERVILLAIN` with a player, the player would
keep the mark ownership and take them to their new role.

```lua
local mvObject = self:AddMarkerVision("crystal_supervillain")
mvObject:SetOwner(ply)
mvObject:SetVisibleFor(VISIBLE_FOR_ROLE)
mvObject:SyncToClients()
```

This would still keep it to the role (the role the player had when
registering the mark). But if they'd update their role to become an
innocent, every innocent would then be able to see it.

Take a look at the entity diffs to see what the API changes mean in
reality:

https://github.com/TTT-2/TTT2/pull/1336/files#diff-25ef72df9ab41eb109b09f41c0d8612a92b99b77e1501762366ca1b09be33e27