Playersettings: Fixed unwanted shared rederence (#1393)
This PR fixes an issue with the synced player settings. Initializing the
synced player settings as `plymeta.playerSettings =
plymeta.playerSettings or {}` results in one table reference being
created that is shared between all instances of the `Player` class. This
pullrequest changes this so that each player has their own instance of
said class.