Fonts: Fix wrong setup (#1378)
Based on @EntranceJew's work I created this new PR that fixes the wrong
font initialization. Due to issues with his branch that I was unable to
resolve, I created this new branch and cherry picked his commit.

While fixing the Font in the DNA scanner I noticed that it was
initialized in its init function meaning that every spawn of a DNA
scanner ent would create a new font. Therefore I moved that part of the
code.

---------

Co-authored-by: EntranceJew <EntranceJew@gmail.com>
Co-authored-by: ZenBre4ker <72046466+ZenBre4ker@users.noreply.github.com>