Scoreboard: Fixed Icons sometimes missing (#1382)
This PR fixes the scoreboard being stuck open after an icon is missing.
This PR does this in two levels

- adding a safety check to make sure the scoreboard never stays stuck
- fixing the root cause that causes the breakage in the first place

## The problem

Some entities register custom projectiles (like the boomerang) that are
then registered as the inflictor. This breaks our current UI as I
expected that the weapon is always a real TTT weapon. The change in this
PR searches for an entity with the given name class, if no weapon was
found.


![image](https://github.com/TTT-2/TTT2/assets/13639408/f9944de7-8995-44eb-9df8-592fffe236ba)

![image](https://github.com/TTT-2/TTT2/assets/13639408/0ac19610-d5b3-4b27-b27e-5bcae8a14c32)

I'm pretty sure that it fixes the two linked issues.

---------

Co-authored-by: ZenBre4ker <72046466+ZenBre4ker@users.noreply.github.com>