Entity Base: Added throw space check (#1413)
Previously you were always able to throw devices (beacon, radio etc).
This could cause issues when really close to a wall as they would glitch
through it. I added the same check to it, that the weapon dropping also
uses.

While working on that I decreased the min distance to a wall when
dropping a weapon as I noticed that a smaller distance works as well and
improves the weapon pickup in tight corners. My testing also showed no
glitches with this change.

This should fix the issue @mexikoedi was having.