remove(sh_sql): Remove adding columns to existing database table
Before this change it was possible to add columns to an existing
database table through the `sql.CreateSqlTable()` function.

This is a troublesome and undocumented(!) functionality which should not
exist going further.
Troublesome as once a column is added there is no way back. This whole
functionality also sways from the `CreateSqlTable` expectation as
running a `CREATE TABLE` statement on an existing sqlite table fails
normally. (In our case we used this function to ensure a table exists,
hence we now return `true` if the table already exists)

`sql.CreateSqlTable` is now also "soft" deprecated through our code
annoations and therefore in our api-docs.
Handling database table definitions should be done through migrations
from now on.

"Hard" deprecation (`ErrorNoHalt` on function call) should follow once
we have migrated our codebase to use migrations to prevent spamming of
deprecation messages.

Additionally this partially fixes #1418 as one of the offending pragma
statements is removed.