Custom World Models: Fixed rendering for custom world models (#1422)
Fixes #1415 

When weapons with a custom world model are dropped, the renderer tries
to render them but fails because it was unable to find the relevant bone
and therefore exits early. This is not a problem when the entity has an
owner (mostly the player holding that weapon) because the player entity
has the required bone if the weapon is set up properly.

This PR fixes this issue by introducing a fallback to render it in place
when the weapon has no owner by using its own main bone (which can have
any random bone name).

Additionally I noticed the hitboxes of the models were completely off
after I fixed the rendering issue. This is because we apply render
filters clientside to the model that only change its appearnce, not its
actual entity. This is not an issue for custom world or view models, but
it is an issue when it is dropped in the world. I fixed this by
detecting if the weapon is in the hands of a player or not. If it isn't,
no clientside modifications should be applied.

There is one part that has to be discussed though:
```lua
-- if a model has a hand bone, it is probably created as a weapon and should therefore
-- be handled that way. If a model is abused aas a weapon then it doesn't have a handbone
-- and should therefore not be rotated
local boneId = boneEntity:LookupBone("ValveBiped.Bip01_R_Hand")
local hasHandBone = boneId and boneId ~= "__INVALIDBONE__"
```

This part doesn't really make sense to me. In my opinion the correct
check here would be to see if the weapon has a valid owner. If it has
one, then the entity can be model (not the entity!) can be resized.
However this completely breaks the hitboxes of the melon mine. The melon
mine is the only one of these that correctly defines a bone that could
be used as an attachment point when used as a weapon. If I use that bone
(`ValveBiped.Bip01_R_Hand`) then the hitboxes are no longer off, as I
now also apply the clientside modifications to the model, that should
only be applied when in hand.
I'm not sure if this is an issue with the melon model or if there is
something else going on. I'd suggest keeping it the way I implemented
and taking a look at it again in case it breaks something in the future.
For now I tested with all models that I had at hand and it worked fine

---------

Co-authored-by: ZenBre4ker <72046466+ZenBre4ker@users.noreply.github.com>