EquipmentMenu: Added check before function call (#1410)
Last night I discussed some things with @mexikoedi. One thing we noticed
that there are quite a few badly done weapons for TTT that forget to set
the weapon base correctly. While this is bad practice and shouldn't be
done, most of these weapons still work just fine. Adding this check here
makes sure that this populate function is only called if it actually
uses the correct base.

---------

Co-authored-by: Histalek <16392835+Histalek@users.noreply.github.com>