Dynamic FOV: Rewrite (#1404)
Due to problems with my previous implementation on low-tickrate servers,
I had to redo the whole system. It now uses the `CalcView` hook as well.
It should work the same (and better) than before.

Things that I could not test: How does this behave when spectating a
player that scopes etc. I think it is fine, but maybe it isn't. I need
some feedback on this, best from @EntranceJew.