Custom View: Fixed broken 3rd party addons (#1439)
There were quite a few reports that the reworked custom view is breaking
3rd party addons such as super sheep. @EntranceJew and I were unable to
reproduce it. This is when it hit me, it is due to the loading order. On
my testing setup, TTT2 is loaded last, because it is a file system
addon. For people using the workshop version this is not true. If TTT2
is loaded before the supersheep, the CalcView hook breaks the
supersheep.

The super sheep registers its own `CalcView` hook:

![image](https://github.com/TTT-2/TTT2/assets/13639408/e6c13350-4716-4d20-bf2a-575524c07669)

However due to the way hooks work in GMod, the next hook is only called
if the previous one returned `nil`. Due to an oversight in my CalcView
hook implementation, there was no early return for observing non-player
entities. I fixed this. This alone should have fixed the issue at hand
here.

But to make it extra safe, I moved the view calculation to
`GM:CalcView`. This way our hook is always called last, which is what we
want.

---------

Co-authored-by: Histalek <16392835+Histalek@users.noreply.github.com>