Bodysearch: Added safeguard to support undefined title or text (#1440)
![image](https://github.com/TTT-2/TTT2/assets/13639408/19997fc6-b623-4cf5-bec0-26edd98d4272)

There was this bugreport. I wasn't able to track down the root of this
problem, but I'm pretty sure addons try to add things without a title
via the hook (which is fine) breaking the rendering code. So I made sure
that not providing a title won't break the rendering.