Sounds: Added multiple UI sounds (#1460)
After testing #1437 I notices that it feels really weird to not have any
kind of feedback when interacting with entities. When picking up
entities, there already is a sound, but not when opening UIs. Therefore
I added `sound.ConditionalPlay` that plays a sound if the sound type is
enabled in the settings.

I also extended this to body search and UI buttons.

See this demo video: https://www.youtube.com/watch?v=c6ixkuFSYfw

Note: I also pulled in the changes from #1351 and added them to the same
system. They are disabled by default.

Menu to disable sounds by type:

![image](https://github.com/TTT-2/TTT2/assets/13639408/fe57417b-f55f-42cd-952d-c632b4aaab1e)

---------

Co-authored-by: EntranceJew <EntranceJew@gmail.com>