Crosshair: Show for nades (#1464)
A crosshair is now also shown when holding a nade (not when the pin is
pulled). This is done so you still know where to point when trying to
interact with an entity.

I also removed the "select on remove" part, as this is in our weapon
base since the last update.