EquipmentMenu: Fix GetClass error (#1466)
This error was reported by @mexikoedi on Discord:

![image](https://github.com/TTT-2/TTT2/assets/13639408/8b6398bd-e2a2-427f-a6fc-2727b76dbabf)

I fixed this by correctly using `WEPS.GetClass` here instead of
`equipment:GetClass`. While doing this I noticed that items have no
equivalent of `GetClass` and `IsValid` so I added that to the base
class. These changes are not strictly needed for this fix, but imho they
are useful for consistency.