ViewModels: Fix error texture bug (#1468)
Fixes this issue reported by @EntranceJew and @mexikoedi:

![image](https://github.com/TTT-2/TTT2/assets/13639408/25bd7d61-fca3-4f9a-b396-e0918053a147)

![image](https://github.com/TTT-2/TTT2/assets/13639408/6336a791-6ffb-46bb-89ed-67e0baf80855)

The bug seen in the image happens because some weapons set a custom view
model texture. Our view model hider resets this custom texture by
calling `viewModel:SetMaterial("")`. I did not anticipate that weapons
might set custom view models, so I wasn't aware that this is an issue.

![image](https://github.com/TTT-2/TTT2/assets/13639408/30296816-9bb0-402e-b4b1-ba7b8904c970)

![image](https://github.com/TTT-2/TTT2/assets/13639408/0fdad9d7-f3b7-4dfd-9439-3f23d6d9e3a5)

We *have to* reset the view model texture because the view model does
not change on weapon change. So if weapon A has a hidden view model and
we apply the debug texture, the same debug texture is applied to weapon
B making it invisible as well. The first solution of applying the
texture reset if it should not be hidden caused the bug mentioned above.

The solution therefore is to set a trigger, if the debug texture is set
before the view model is rendered and reset the material after the view
model is rendered. If a player then switches to weapon B, the view model
works as expected without a continuous material reset.

---------

Co-authored-by: Histalek <16392835+Histalek@users.noreply.github.com>