Crosshair: Prevent rendering GMod crosshair (#1471)
This started as a fix for a bug introduced in #1464 because HUD Help was
now drawn twice for grenades with HUD help. So I decided to make the
crosshairq function callable, so it can be called seperately in the
grenade base.

While doing so I noticed
[`SWEP:DoDrawCrosshair`](https://wiki.facepunch.com/gmod/WEAPON:DoDrawCrosshair).
This function is called when `SWEP.DrawCrosshair` is set to true. We
have set that to false in our base, because we do not want the GMod
crosshair do be visible at all. However, some weapons set this to true,
probably because the SWEP author thinks setting this to false disables
the TTT2 crosshair. This results in two crosshairs being drawn at once:

![image](https://github.com/TTT-2/TTT2/assets/13639408/ac2fe7b6-0a39-499f-a415-0b9263386efe)

In a perfect world we would use `SWEP:DoDrawCrosshair` to draw our own
crosshair and then return `true` at the end. That way the GMod crosshair
is never drawn and SWEP authors can use `SWEP.DrawCrosshair` to disable
the crosshair for their weapon. However I know for a fact that there are
a few weapons that explicitly set this variable to `false` which would
result in no crosshair being drawn at all, even though this is not the
intention.

My solution here was to use this hook to draw the crosshair and pass
another variable to it. If it is called by us, the crosshair should be
drawn. If it is called by the engine, the hook should just return true.
This way all weapons always have a crosshair and only the TTT2
crosshair.

We could also opt to use `SWEP:DoDrawCrosshair` in its intended way
hoping that those authors fix their weapons, but I think that this would
cause people complaining about TTT2.