EquipmentEditor: Set clip on buy (#1499)
Added `SWEP.SetClipOnBuy` and `SWEP.ClipOnBuy` to set the weapon's clip
on buy via the equipment editor. I did not want to just use the latter
variable to set a clip on buy to all weapons as most "one time use
weapons" instantly remove the weapon on use which makes this slider
pointless. It is therefore something swep makers have to integrate into
their weapon, which is pretty simple.

I also reordered some elements in the equipment menu because they were
out of place. It now looks like this:


![image](https://github.com/TTT-2/TTT2/assets/13639408/3c0a6722-1ee1-4c23-b1a5-a29cdd2264fb)

---------

Co-authored-by: Histalek <16392835+Histalek@users.noreply.github.com>
Co-authored-by: ZenBre4ker <72046466+ZenBre4ker@users.noreply.github.com>