HUD: Fixed OHI Desync (#1513)
Sometimes overhead icons stayed stuck in the world. This is due to the
way GMod/source handles player positions. If a player is far enough
away, their position is no longer updated (as can be seen with wallhacks
for example).

While `:GetPos()` still returns a position that at least represents a
rough estimate of their position, the bone position stays at the last
updated place. This means that the overhead icon stays floating in
random places.

This pullrequest fixes this by introducing a check that uses the
fallback position if the player position and the bone position diverge
too much.

While I tested this, it would probably be good if @mexikoedi could
confirm this as well.

---------

Co-authored-by: Histalek <16392835+Histalek@users.noreply.github.com>