Voice: Reworked HUD (#1473)
This doesn't do much besides creating a new UI in our HUD system instead
of the weird stuff that was done before. It also adds a function that
generates a fake voice spectrum.


![image](https://github.com/TTT-2/TTT2/assets/13639408/b7c56b91-0dbe-4fd5-9b19-a18501eefd11)

Note: This voice spectrum is only visible for other players. Or if you
have set `voice_loopback 1` on the client. I'm not sure if there is any
real solution to that, so I guess it is unfixable.

Stuff I want to add to voice (in other PRs):
- indicator in the UI if this voice is locational or global, maybe a
neat detective item would be for him to be able to talk globally, even
if locational voice is turned on
- right now there exists only push to talk, I'd like to add push to mute
and auto detection as well
- a neat UI for the battery drain state (that's something I removed in
this PR, but I will add it back. I'm not sure if anyone is using that
anyway)
- clean up that cl_voice stuff, it looks like a mess

For the reviewers: Besides the new UI elements not much changed. I
reordered some calls, renamed some variables (`clrd`, seriously? :D) and
removed the old rendering code.

---------

Co-authored-by: ruby0b <106119328+ruby0b@users.noreply.github.com>