F1 menu: Added Option to Change Maps via the F1 Menu (#1512)
This adds the option for admins to change the currently played map while
also starting an admin commands menu. This makes use of the combo cards
added in #1511.


![image](https://github.com/TTT-2/TTT2/assets/13639408/15eb3fdf-e471-4610-a9b3-b56e97bc424b)


![image](https://github.com/TTT-2/TTT2/assets/13639408/ce814f76-aa30-4424-80d7-d4a40870ecd0)


![image](https://github.com/TTT-2/TTT2/assets/13639408/cb6ce1ad-d113-406c-b91a-1578c827cf9f)

One thing that was added, is a hook that is called, once a client
finished the hot reload. This is needed so that the server knows when
the client is ready to resend the data:
```lua
---
-- Called when a player is ready again after they reloaded their game. Can be used to send
-- data that is needed after a reload.
-- @param Player ply The player that finished the reloading
-- @hook
-- @realm server
function GM:TTT2PlayerFinishedReloading(ply)
```

This PR also greatly extended the `map` module which can also then be
used by other addons. Details can be checked in the changed files, but
most notably are these additions:

This function returns a cached material for a given map. This also
includes maps that don't have a material. Since this works by
downloading these materials from steam, this system works as long as the
map is installed via the workshop and the workshop page has a preview
image.
```lua
---
-- Returns the cached icon of a map if there is an icon available.
-- @note This not only uses the icons available on the client through addons, it also
-- searches the workshop and tries to assign the corrent workshop icon if there
-- is no locally available map icon.
-- @param string name The map name
-- @return nil|IMaterial Returns the cached material if available
-- @realm client
function map.GetIcon(name)
```

This function returns a list of all available maps. It already excludes
those maps, where the prefix was disabled in the admin settings. This
can be used in map mote addons as a map base.
```lua
---
-- Returns a sanitized list of maps installed on the server. Sanitized means that only
-- valid maps with a prefix that is enabled are listed here. This list is automatically
-- synced between the server and all clients.
-- @return table Returns a table with the map names
-- @realm shared
function map.GetList()
```