Loadingscreen: Added Screen on Round Restart (#1524)
**This PR is not as big as it looks, this is only because I modified the
tips strings in all languages!**

This pullrequests adds a loadingscreen that hides the visible and
audible lag introduced by the map cleanup on round change. This is done
because the lag can hardly be improved since it all takes place in
`game.CleanupMap`. It can be disabled if one dislikes it.

This PR also moves the (in my opinion annoying) tips panel to a more
useful place by displaying those tips on the new loading screen.


https://github.com/TTT-2/TTT2/assets/13639408/239e0441-0d51-4868-834d-ec121443e86b

Settings panel to disable loading screen and tips:


![image](https://github.com/TTT-2/TTT2/assets/13639408/9b25cacf-e325-4700-8606-8e37165b6285)

Note: The tips convar is carried over from the original TTT convar.