Voice Chat: Added module for battery drain and bar to UI (#1536)
The voice battery drain was previously removed from the UI in the voice
rework. It is now added back to the UI. While doing so I moved all
convars and functions to a new module called `voicebattery`. That way I
was able to remove the global bools and replace them with with
replicated convars.


![image](https://github.com/TTT-2/TTT2/assets/13639408/5052f54c-11b1-4197-b6f1-0678951fb45f)

I'm not sure that I'm a fan of the voice battery being client only. This
seems like a bad idea to me. But I wanted to keep it that way for now,
we might want to redo it at some point.