Scoreboard/Specs: Changed Groups for Scoreboard Search (#1545)
Fixes #1483

When a player is dead and confirmed, you can see the search info in the
scoreboard when clicking on their name. This did not work for players
that died because they forced themselves as a spectator. This is because
of this `if/else` structure that did not handle `GROUP_SPEC`.

As there only exists these four groups:
```lua
GROUP_TERROR = 1
GROUP_NOTFOUND = 2
GROUP_FOUND = 3
GROUP_SPEC = 4
```

I decided to just changes this to an `else`. The only sideeffect is that
you are now able to click on any spectator in the scoreboard and it says
that this player has no search information available. To me this seems
like a valid solution.