ENTITY: New Use workflow (#1458)
This is implementing the new use workflow and gets rid of UseOverride.
Also C4 is now the first to properly use ClientUse to display a UI
without routing it over the server.
As sidenote I fixed the Visualizer so that it can be picked up by the
originator. This was only a problem if a non policing role should be
able to buy it. Thats more in line with our general shop setup
possibilities.

---------

Co-authored-by: Tim Goll <github@timgoll.de>